권한관리 기능 구현

+++ src/main/java/hello/hellospring/service/PostService.java
@@ -1,6 +1,7 @@
 package hello.hellospring.service;
 
 
+import hello.hellospring.common.exception.UnAuthorizationException;
 import hello.hellospring.domain.Member;
 import hello.hellospring.domain.Post;
 import hello.hellospring.dto.PostDto;
@@ -18,6 +19,7 @@ import java.util.NoSuchElementException;
 @RequiredArgsConstructor
 public class PostService {
     private final PostRepository postRepository;
+    private final LoginService loginService;
 
     public void saveNewPost(PostDto.Request request, Member member) {
         Post post = request.toEntity(member);
@@ -25,7 +27,9 @@ public class PostService {
     }
 
     public void updatePost(Post post, PostDto.Request request) {
-        post.update(request);
+        if (isOwner(post)) {
+            post.update(request);
+        }
     }
 
     @Transactional(readOnly = true)
@@ -36,11 +40,22 @@ public class PostService {
     }
 
     public void deletePost(Long id) {
-        postRepository.deleteById(id);
+        Post post = postRepository.findById(id).orElseThrow(NoSuchElementException::new);
+        if (isOwner(post)) {
+            postRepository.deleteById(id);
+        }
     }
 
     @Transactional(readOnly = true)
     public Page<Post> findAllPosts(Pageable pageable) {
         return postRepository.findAll(pageable);
+    }
+
+    public boolean isOwner(Post post) {
+        Member member = loginService.getLoggedInMember();
+        if (member != post.getMember()) {
+            throw new UnAuthorizationException();
+        }
+        return true;
     }
 }
